% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anderson_Darling_test.R
\name{Anderson_Darling_test}
\alias{Anderson_Darling_test}
\title{Anderson-Darling Normality Test}
\usage{
Anderson_Darling_test(
  x,
  alpha = 0.05,
  silent = FALSE,
  summary = TRUE,
  misc = FALSE
)
}
\arguments{
\item{x}{A numeric vector, at least length of 8.}

\item{alpha}{Numeric (default: 0.05). Significance threshold, range from 0 to 1.}

\item{silent}{Logical (default: FALSE). If \code{FALSE}, print out the results.}

\item{summary}{Logical (default: TRUE). Produce a summary table.}

\item{misc}{Logical (default: FALSE). Output other unimportant parameters.}
}
\value{
A list.
}
\description{
Performs the Anderson–Darling (A\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}) normality test, an EDF-based
goodness-of-fit test that gives greater weight to deviations in the tails
of the distribution.
}
\examples{
out <- Anderson_Darling_test(rnorm(10))
print(out$summary)

}
\references{
D’Agostino, R.B., 2017. Tests for the Normal Distribution.
In: D'Agostino, R.B., Stephens, M.A. (Eds.),
Goodness-of-Fit Techniques, 1st ed. Routledge, New York,
pg. 372–373.
https://doi.org/10.1201/9780203753064

Stephens, M.A., 2017. Tests Based on EDF Statistics.
In: D'Agostino, R.B., Stephens, M.A. (Eds.),
Goodness-of-Fit Techniques, 1st ed. Routledge, New York,
(Table 4.8 & 4.9) pg. 126–128.
https://doi.org/10.1201/9780203753064

Anderson, T.W., Darling, D.A., 1954.
A Test of Goodness of Fit.
J. Am. Stat. Assoc. 49, 765–769.
https://doi.org/10.1080/01621459.1954.10501232
}
