% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholesterol}
\alias{cholesterol}
\title{Cholesterol data}
\format{
A numeric vector length of 62.
}
\usage{
cholesterol
}
\description{
A numeric vector, the cholesterol values from a sample of 62 subjects from the
Framingham Heart Study (FHS). This dataset was obtained from D'Agostino paper.
}
\references{
D’Agostino, R.B., Belanger, A., D’Agostino Jr., R.B., 1990.
A Suggestion for Using Powerful and Informative Tests of Normality.
Am. Stat. 44, 316–321.
https://doi.org/10.1080/00031305.1990.10475751
}
\keyword{datasets}
