% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cramer_von_Mises_test.R
\name{Cramer_von_Mises_test}
\alias{Cramer_von_Mises_test}
\title{Cramer-von Mises Normality Test}
\usage{
Cramer_von_Mises_test(
  x,
  alpha = 0.05,
  silent = FALSE,
  summary = TRUE,
  misc = FALSE
)
}
\arguments{
\item{x}{A numeric vector, at least length of 8.}

\item{alpha}{Numeric (default: 0.05). Significance threshold, range from 0 to 1.}

\item{silent}{Logical (default: FALSE). If \code{FALSE}, print out the results.}

\item{summary}{Logical (default: TRUE). Produce a summary table.}

\item{misc}{Logical (default: FALSE). Output other unimportant parameters.}
}
\value{
A list.
}
\description{
An empirical distribution function (EDF)-based goodness-of-fit test that measures the
overall discrepancy between the empirical and theoretical cumulative distribution functions
by assigning relatively uniform weight across the entire distribution.
}
\examples{
out <- Cramer_von_Mises_test(rnorm(10))
print(out$summary)

}
\references{
Thode, H. C., Jr., 2002. Goodness of fit tests.
Testing for normality. Marcel Dekker, New York.
(Section 5.1.3, pg. 103)

Stephens, M.A., 2017. Tests Based on EDF Statistics.
In: D'Agostino, R.B., Stephens, M.A. (Eds.),
Goodness-of-Fit Techniques, 1st ed. Routledge, New York,
(Table 4.9, pg. 127).
https://doi.org/10.1201/9780203753064
}
