% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.Agostino_Pearson_test.R
\name{D.Agostino_Pearson_test}
\alias{D.Agostino_Pearson_test}
\title{D'Agostino–Pearson K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}} Normality Test}
\usage{
D.Agostino_Pearson_test(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  silent = FALSE,
  summary = TRUE,
  misc = FALSE
)
}
\arguments{
\item{x}{Numeric vector. Must have length at least 20.}

\item{alpha}{Numeric (default: 0.05). Significance level for hypothesis testing.
Must be between 0 and 1.}

\item{alternative}{Character (default: "two.sided").
Specifies the alternative hypothesis. Available options are
c("two.sided", "less", "greater"). Note that this option is only applied
to the skewness and kurtosis components of the test.}

\item{silent}{Logical (default: FALSE). If \code{FALSE}, results are printed
to the console.}

\item{summary}{Logical (default: TRUE). Produce a summary table.}

\item{misc}{Logical (default: FALSE). Output other unimportant parameters.}
}
\value{
A list
}
\description{
The D'Agostino–Pearson chi-squared (K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}) test is a moment-based
omnibus test for normality.
}
\details{
It evaluates the null hypothesis that the data come from a normal distribution
by combining standardized measures of skewness and kurtosis into a single
chi-squared test statistic.
}
\examples{
out <- D.Agostino_Pearson_test(rnorm(50))
print(out$summary)

}
\references{
D’Agostino, R.B., Belanger, A., D’Agostino, R.B., 1990.
A Suggestion for Using Powerful and Informative Tests of Normality.
Am. Stat. 44, 316–321.
https://doi.org/10.1080/00031305.1990.10475751
}
