% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapiro_Wilk_test.R
\name{.Shapiro_Wilk_original}
\alias{.Shapiro_Wilk_original}
\title{Shapiro-Wilk Normality Test}
\usage{
.Shapiro_Wilk_original(x, alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector.}

\item{alpha}{Numeric (default: 0.05). Significance threshold, range from 0 to 1.}
}
\value{
A list, see \code{normality_standard_output()}.
}
\description{
The original version of the W test.
}
\references{
Shapiro, S.S., Wilk, M.B., 1965.
An Analysis of Variance Test for Normality (Complete Samples).
Biometrika 52, 591–611.
https://doi.org/10.2307/2333709
}
