% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tied}
\alias{is_tied}
\title{Tied data}
\usage{
is_tied(x, ratio = 0.3, remove_NA = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{ratio}{Numeric (default: 0.3).
The ratio threshold of being considred as tied-data. The value range from 0 to 1.}

\item{remove_NA}{Logical (default: TRUE). Whether or not to remove NAs.}
}
\value{
Logical
}
\description{
Tied data
}
\examples{
is_tied(c(1, 1, 2, 2, 2, 3, 4, 5))
#> TRUE
}
