% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis test}
\usage{
kurtosis(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  method = c("G2", "b2", "g2")
)
}
\arguments{
\item{x}{Numeric vector. The input data.}

\item{alpha}{Numeric (default: 0.05). Significance threshold (0 - 1).}

\item{alternative}{Character (default: "two.sided).
The alternative hypothesis (H1) to test. Available options are c("two.sided", "less", "greater").}

\item{method}{Character (default: "G2"). Different skewness formula.
Available options are c("G2", "b2", "g2"). The "g2" is the original one.
The "G2" and "b2" are the unbiased estimate version of "g2".}
}
\value{
A list:
is_normal: Is the input data normally distributed?
method: The name of the test.
alpha: Significance threshold (default: 0.05).
alternative: The alternative hypothesis (H1) to test.
summary_table: Statistic summary, if any.
statistic: The value used to calculate p-value.
pvalue: p-value.
confidence_interval: The lower and upper bound of CI.
}
\description{
Kurtosis test
}
\examples{
d1 <- c(10:17, 12, 12, 13, 13, 13, 13, 13, 14, 14, 14, 15, 15)
kurtosis(d1)
}
\references{
Joanes, D.N., Gill, C.A., 1998.
Comparing measures of sample skewness and kurtosis.
J Royal Statistical Soc D 47, 183–189.
https://doi.org/10.1111/1467-9884.00122

Wright, D.B., Herrington, J.A., 2011.
Problematic standard errors and confidence intervals for skewness and kurtosis.
Behav Res 43, 8–17.
https://doi.org/10.3758/s13428-010-0044-x
}
