% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leghorn_chick}
\alias{leghorn_chick}
\title{Leghorn chicken data}
\format{
A numeric vector length of 20.
}
\usage{
leghorn_chick
}
\description{
A numeric vector
}
\references{
Stephens, M.A., 2017. Tests Based on EDF Statistics.
In: D’Agostino, R.B., Stephens, M.A. (Eds.),
Goodness-of-Fit Techniques, 1st ed. Routledge, New York,
pp. 98.
https://doi.org/10.1201/9780203753064
}
\keyword{datasets}
