% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lilliefors_test.R
\name{Lilliefors_test}
\alias{Lilliefors_test}
\title{Lilliefors Normality Test}
\usage{
Lilliefors_test(x, alpha = 0.05, silent = FALSE, summary = TRUE, misc = FALSE)
}
\arguments{
\item{x}{A numeric vector, at least length of 8.}

\item{alpha}{Numeric (default: 0.05). Significance threshold, range from 0 to 1.}

\item{silent}{Logical (default: FALSE). If \code{FALSE}, print out the results.}

\item{summary}{Logical (default: TRUE). Produce a summary table.}

\item{misc}{Logical (default: FALSE). Output other unimportant parameters.}
}
\value{
A list.
}
\description{
Performs the Lilliefors normality test, which is an empirical distribution
function (EDF)-based goodness-of-fit test derived from the Kolmogorov–Smirnov
test, using the approximation proposed by Molin and Abdi (1998).
}
\examples{
out <- Lilliefors_test(rnorm(10))
print(out$summary)

}
\references{
Molin, P., Abdi, H., 1998.
New tables and numerical approximation for the
Kolmogorov-Smirnov/Lillierfors/Van Soest test of normality.
Technical report, University of Bourgogne.
}
