% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_standard-output.R
\name{normality_standard_output}
\alias{normality_standard_output}
\title{Standard output format}
\usage{
normality_standard_output(
  method = "what test?",
  bool = NA,
  alpha = NA_real_,
  alternative = c("two.sided", "less", "greater"),
  summary_table = NULL,
  statistic = NA_real_,
  pvalue = NA_real_,
  confidence_interval = c(lower = NA_real_, upper = NA_real_)
)
}
\arguments{
\item{method}{Character. The name of the test.}

\item{bool}{Logical. Is the input data normally distributed?}

\item{alpha}{Numeric (default: 0.05). Significance threshold.}

\item{alternative}{Character. The alternative hypothesis (H1) to test.
Available options are c("two.sided", "less", "greater").}

\item{summary_table}{Statistic summary, if any.}

\item{statistic}{Numeric. The value used to calculate p-value.}

\item{pvalue}{Numeric. The p-value of the test.}

\item{confidence_interval}{Numeric vector of length 2. The lower and upper bound of CI.}
}
\value{
A list contains 8 vectors.
}
\description{
The standard output format for \code{normality} package.
}
