% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shapiro_Wilk_coef_table}
\alias{Shapiro_Wilk_coef_table}
\title{Shapiro-Wilk normality test (coefficients)}
\format{
A data frame with 50 rows and 25 variables:
\describe{
rownames is the sample size (n);
colnames is the corresponding coefficients (a\if{html}{\out{<sub>}}i\if{html}{\out{</sub>}}).
}
}
\usage{
Shapiro_Wilk_coef_table
}
\description{
Coefficients (a\if{html}{\out{<sub>}}i\if{html}{\out{</sub>}}) for the W test for normality.
}
\references{
Shapiro, S.S., Wilk, M.B., 1965.
An Analysis of Variance Test for Normality (Complete Samples).
Biometrika 52, 591–611.
https://doi.org/10.2307/2333709
}
\keyword{datasets}
