% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shapiro_Wilk_pval_table}
\alias{Shapiro_Wilk_pval_table}
\title{Shapiro-Wilk normality test (p-values)}
\format{
A data frame with 50 rows and 10 variables:
\describe{
rownames is the sample size (n);
colnames is the corresponding p-values.
}
}
\usage{
Shapiro_Wilk_pval_table
}
\description{
The percentage points (critical values of W) of the W test for n = 3(1)50.
}
\references{
Shapiro, S.S., Wilk, M.B., 1965.
An Analysis of Variance Test for Normality (Complete Samples).
Biometrika 52, 591–611.
https://doi.org/10.2307/2333709
}
\keyword{datasets}
