% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapiro_Wilk_test.R
\name{Shapiro_Wilk_test}
\alias{Shapiro_Wilk_test}
\title{Shapiro-Wilk Normality Test}
\usage{
Shapiro_Wilk_test(x, alpha = 0.05, method = c("swr", "sf", "sw"))
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{Significance threshold (default: 0.05).}

\item{method}{Character (default: "swr"). Use which modification of the test?
Available options are c("swr", "sf", "sw").}
}
\value{
A list.
}
\description{
Performs the Shapiro-Wilk test of normality.
}
\details{
method
\itemize{
\item "swr": Shapiro-Wilk-Royston, extended by Royston.
\item "sf": Shapiro-Francia, modified by Francia.
\item "sw": Shapiro-Wilk, the original test.
}
}
\examples{
Shapiro_Wilk_test(rnorm(20), method = "sw")
}
