% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anderson_Darling_test.R
\name{Anderson_Darling_test}
\alias{Anderson_Darling_test}
\title{Anderson-Darling Normality Test}
\usage{
Anderson_Darling_test(x, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{Numeric (default: 0.05). Significance threshold, range from 0 to 1.}

\item{verbose}{Logical (default: FALSE). Show messages.}
}
\value{
A list:
\itemize{
\item is_normal: Is the input data normally distributed?
\item method: The name of the test.
\item alpha: Significance threshold (default: 0.05).
\item alternative: The alternative hypothesis (H1) to test.
\item summary_table: Statistic summary, if any. Mostly output as a data frame.
\item statistic: The value used to calculate p-value.
\item pvalue: The \if{html}{\out{<i>}}p\if{html}{\out{</i>}} value.
\item confidence_interval: The lower and upper bound of confidence interval (CI).
}
}
\description{
Anderson-Darling Normality Test
}
\examples{
Anderson_Darling_test(leghorn_chick)

}
\references{
D’Agostino, RalphB., 2017.
Goodness-of-Fit Techniques, page 123-128 & 372-373. 1st ed. Routledge.
https://doi.org/10.1201/9780203753064
}
