% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAgostino-Pearson_test.R
\name{D.Agostino_Pearson_test}
\alias{D.Agostino_Pearson_test}
\title{D'Agostino-Pearson K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}} Normality Test}
\usage{
D.Agostino_Pearson_test(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  min_n = 20
)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{Significance threshold (default: 0.05).}

\item{alternative}{Character (default: "two.sided).
The alternative hypothesis (H1) to test. Available options are c("two.sided", "less", "greater").
Note that, this is only applied on skewness and kurtosis test.}

\item{min_n}{Integer. The minimum observations required (default: 20).}
}
\value{
A list:
\itemize{
\item is_normal: Is the input data normally distributed?
\item method: The name of the test.
\item alpha: Significance threshold (default: 0.05).
\item alternative: The alternative hypothesis (H1) to test.
\item summary_table: Statistic summary, if any. Mostly output as a data frame.
\item statistic: The value used to calculate p-value.
\item pvalue: The \if{html}{\out{<i>}}p\if{html}{\out{</i>}} value.
\item confidence_interval: The lower and upper bound of confidence interval (CI).
}
}
\description{
The D'Agostino–Pearson Chi-square (K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}) test is a statistical test for
assessing whether a sample comes from a normal distribution.
It combines information from skewness (asymmetry) and kurtosis (tail heaviness)
into a single omnibus test statistic.
}
\examples{
D.Agostino_Pearson_test(cholesterol)

}
\references{
D’agostino, R.B., Belanger, A., D’agostino, R.B., 1990.
A Suggestion for Using Powerful and Informative Tests of Normality.
The American Statistician 44, 316–321.
https://doi.org/10.1080/00031305.1990.10475751
}
