% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAgostino-Pearson_test.R
\name{D.Agostino_Pearson_test}
\alias{D.Agostino_Pearson_test}
\title{D'Agostino-Pearson K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}} Normality Test}
\usage{
D.Agostino_Pearson_test(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  silent = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{Significance threshold (default: 0.05).}

\item{alternative}{Character (default: "two.sided).
The alternative hypothesis (H1) to test. Available options are c("two.sided", "less", "greater").
Note that, this is only applied on skewness and kurtosis test.}

\item{silent}{Logical (default: FALSE). If \code{FALSE}, print out the results.}
}
\value{
A list
}
\description{
The D'Agostino–Pearson Chi-square (K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}) test is a moment test for
assessing whether a sample comes from a normal distribution.
It combines information from skewness (asymmetry) and kurtosis (tail heaviness)
into a single omnibus test statistic.
}
\examples{
D.Agostino_Pearson_test(cholesterol)

}
\references{
D’Agostino, R.B., Belanger, A., D’Agostino, R.B., 1990.
A Suggestion for Using Powerful and Informative Tests of Normality.
Am. Stat. 44, 316–321.
https://doi.org/10.1080/00031305.1990.10475751
}
