% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAgostino-Pearson_test.R
\name{D.Agostino_Pearson_test}
\alias{D.Agostino_Pearson_test}
\title{D'Agostino-Pearson chi square test of data normality.}
\usage{
D.Agostino_Pearson_test(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  min_n = 20
)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{Significance threshold (default: 0.05).}

\item{alternative}{Character (default: "two.sided).
The alternative hypothesis (H1) to test. Available options are c("two.sided", "less", "greater").
Note that, this is only applicable on skewness and kurtosis test, not functioning on the omnibus K-square test.}

\item{min_n}{Integer. The minimum observations required (default: 20).}
}
\value{
A list:
is_normal: Is the input data normally distributed?
method: The name of the test.
alpha: Significance threshold (default: 0.05).
alternative: The alternative hypothesis (H1) to test.
summary_table: Statistic summary, if any.
statistic: The value used to calculate p-value.
pvalue: p-value.
confidence_interval: The lower and upper bound of CI.
}
\description{
The D'Agostino–Pearson chi square (K\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}) test is a statistical test for assessing whether a sample comes from a normal distribution.
It combines information from:
\itemize{
\item skewness (asymmetry)
\item kurtosis (tail heaviness)
into a single omnibus test statistic.
}
}
\examples{
D.Agostino_Pearson_test(cholesterol)

}
\references{
D’agostino, R.B., Belanger, A., D’agostino, R.B., 1990.
A Suggestion for Using Powerful and Informative Tests of Normality.
The American Statistician 44, 316–321.
https://doi.org/10.1080/00031305.1990.10475751
}
