% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapiro_Wilk_test.R
\name{.Shapiro_Francia}
\alias{.Shapiro_Francia}
\title{Shapiro-Francia Normality Test}
\usage{
.Shapiro_Francia(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A list.
}
\description{
Performs the Shapiro-Francia test of normality.
}
\references{
Shapiro, S.S., Francia, R.S., 1972.
An Approximate Analysis of Variance Test for Normality.
J. Am. Stat. Assoc. 67, 215–216.
https://doi.org/10.1080/01621459.1972.10481232

Royston, P., 1993.
A pocket-calculator algorithm for the Shapiro–Francia test for non-normality:
an application to medicine.
Stat. Med. 12, 181–184.
https://doi.org/10.1002/sim.4780120209
}
