% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_normal.R
\name{is_normal}
\alias{is_normal}
\title{Normality test}
\usage{
is_normal(data, formula = NULL, alpha = 0.05, sensitivity = 2, summary = FALSE)
}
\arguments{
\item{data}{A data frame or a numeric vector.}

\item{formula}{Formula (default: NULL).
If \code{data} is a data frame, define the val ~ group.}

\item{alpha}{Significance threshold, range from 0 to 1 (default: 0.05).}

\item{sensitivity}{Numeric, range from 1 to 3 (default: 2).
The greater the value, the greater chance to consider as non-normal.}

\item{summary}{Logical (default: FALSE). If \code{TRUE}, show the summary table.}
}
\value{
A list or a boolean.
}
\description{
A handy wrapper for data normality test using the Shapiro-Wilk-Royston,
D'Agostino-Pearson, and Anderson-Darling tests.
}
\examples{
is_normal(iris, Sepal.Length ~ Species)
}
