% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leghorn_chick}
\alias{leghorn_chick}
\title{Leghorn chick data}
\format{
A numeric vector length of 20.
}
\usage{
leghorn_chick
}
\description{
A numeric vector
}
\references{
D’Agostino, RalphB., 2017.
Goodness-of-Fit Techniques, page 98 (Table 4.1). 1st ed. Routledge.
https://doi.org/10.1201/9780203753064
}
\keyword{datasets}
