% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{normality}
\alias{normality}
\title{Normality: 17 standard univariate normality tests in a single wrapper.}
\usage{
normality(values)
}
\arguments{
\item{values}{A numerical vector of values}
}
\value{
A tibble containing the results of performing the following normality tests:
\itemize{
\item Adjusted Jarque-Bera
\item Anderson-Darling
\item Anscombe-Glynn
\item Bonett-Seier
\item D'Agostino skewness
\item Frosini
\item Geary
\item Hegazy-Green
\item Jarque-Bera
\item Lilliefors
\item Minimum Hellinger Distance
\item Pearson chi-square
\item Shapiro-Francia
\item Shapiro-Wilk
\item Spiegelhalter
\item Weisberg-Bingham
}
}
\description{
Normality: 17 standard univariate normality tests in a single wrapper.
}
\details{
The package is just a simple wrapper to the \code{moments}, \code{normtest}, \code{nortest}, \code{stats} and \code{mhde} packages where these functions are implemented.
}
\examples{
# Test normal random deviates:
values <- rnorm(100)
normality(values)
}
