% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normality_standard_output}
\alias{normality_standard_output}
\title{Standard output format}
\usage{
normality_standard_output(
  method = "what test?",
  is_normal = NA,
  alpha = NA_real_,
  alternative = c("two.sided", "less", "greater"),
  summary = NULL,
  statistic = NA_real_,
  pvalue = NA_real_
)
}
\arguments{
\item{method}{Character. The name of the test.}

\item{is_normal}{Logical. Is the input data normally distributed?}

\item{alpha}{Numeric (default: 0.05). Significance threshold.}

\item{alternative}{Character. The alternative hypothesis (H1) to test.
Available options are c("two.sided", "less", "greater").}

\item{summary}{Statistic summary, if any.}

\item{statistic}{Numeric. The value used to calculate p-value.}

\item{pvalue}{Numeric. The p-value of the test.}
}
\value{
A list.
}
\description{
The standard output format for \code{normality} package.
}
