% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness test}
\usage{
skewness(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  method = c("G1", "b1", "g1"),
  silent = FALSE,
  summary = TRUE
)
}
\arguments{
\item{x}{Numeric vector containing the input data.}

\item{alpha}{Numeric (default: 0.05). Significance level for hypothesis
testing. Must be between 0 and 1.}

\item{alternative}{Character (default: "two.sided").
Specifies the alternative hypothesis. Available options are
c("two.sided", "less", "greater").}

\item{method}{Character (default: "G1"). Formula used to estimate
skewness. Available options are c("G1", "b1", "g1").
The "g1" statistic is the conventional moment-based sample skewness.
The "G1" and "b1" statistics apply finite-sample corrections to reduce
the bias of "g1".}

\item{silent}{Logical (default: FALSE). If \code{FALSE}, the test results are
printed to the console.}

\item{summary}{Logical (default: TRUE). Produce a summary table.}
}
\value{
A list
}
\description{
The test evaluates whether the population skewness is equal to zero.
Under the null hypothesis, the data are assumed to originate from a
symmetric distribution. Significant positive or negative skewness
indicates asymmetry in the distribution and may suggest a departure
from normality.
}
\examples{
skewness(rnorm(30))
}
\references{
Joanes, D.N., Gill, C.A., 1998.
Comparing measures of sample skewness and kurtosis.
J. R. Stat. Soc. D (The Statistician) 47, 183–189.
https://doi.org/10.1111/1467-9884.00122

Wright, D.B., Herrington, J.A., 2011.
Problematic standard errors and confidence intervals for skewness and kurtosis.
Behav. Res. Methods 43, 8–17.
https://doi.org/10.3758/s13428-010-0044-x
}
